add file examples
